+++ test/Benchmarks/Trace/TraceBenchmarks.cs
@@ -31,60 +31,92 @@ Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
 
 namespace Benchmarks.Trace;
 
+#pragma warning disable CA1001 // Types that own disposable fields should be disposable - handled by GlobalCleanup
 public class TraceBenchmarks
+#pragma warning restore CA1001 // Types that own disposable fields should be disposable - handled by GlobalCleanup
 {
-    private readonly ActivitySource sourceWithNoListener = new("Benchmark.NoListener");
-    private readonly ActivitySource sourceWithPropagationDataListner = new("Benchmark.PropagationDataListner");
-    private readonly ActivitySource sourceWithAllDataListner = new("Benchmark.AllDataListner");
-    private readonly ActivitySource sourceWithAllDataAndRecordedListner = new("Benchmark.AllDataAndRecordedListner");
-    private readonly ActivitySource sourceWithOneProcessor = new("Benchmark.OneProcessor");
-    private readonly ActivitySource sourceWithTwoProcessors = new("Benchmark.TwoProcessors");
-    private readonly ActivitySource sourceWithThreeProcessors = new("Benchmark.ThreeProcessors");
-    private readonly ActivitySource sourceWithOneLegacyActivityOperationNameSubscription = new("Benchmark.OneInstrumentation");
-    private readonly ActivitySource sourceWithTwoLegacyActivityOperationNameSubscriptions = new("Benchmark.TwoInstrumentations");
+    private ActivitySource? sourceWithNoListener;
+    private ActivitySource? sourceWithPropagationDataListner;
+    private ActivitySource? sourceWithAllDataListner;
+    private ActivitySource? sourceWithAllDataAndRecordedListner;
+    private ActivitySource? sourceWithOneProcessor;
+    private ActivitySource? sourceWithTwoProcessors;
+    private ActivitySource? sourceWithThreeProcessors;
+    private ActivitySource? sourceWithOneLegacyActivityOperationNameSubscription;
+    private ActivitySource? sourceWithTwoLegacyActivityOperationNameSubscriptions;
 
-    public TraceBenchmarks()
+    private TracerProvider? tracerProvierWithOneProcessor;
+    private TracerProvider? tracerProvierWithTwoProcessors;
+    private TracerProvider? tracerProvierWithThreeProcessors;
+    private TracerProvider? tracerProvierWithOneLegacyActivityOperationNameSubscription;
+    private TracerProvider? tracerProvierWithTwoLegacyActivityOperationNameSubscriptions;
+    private TracerProvider? tracerProvierWithExactMatchLegacyActivityListner;
+    private TracerProvider? tracerProvierWithWildcardMatchLegacyActivityListner;
+
+    private ActivityListener? activityListenerPropagationData;
+    private ActivityListener? activityListenerAllData;
+    private ActivityListener? activityListenerAllDataAndRecordedData;
+
+    [GlobalSetup]
+    public void Setup()
     {
         Activity.DefaultIdFormat = ActivityIdFormat.W3C;
 
-        ActivitySource.AddActivityListener(new ActivityListener
+        this.sourceWithNoListener = new("Benchmark.NoListener");
+        this.sourceWithPropagationDataListner = new("Benchmark.PropagationDataListner");
+        this.sourceWithAllDataListner = new("Benchmark.AllDataListner");
+        this.sourceWithAllDataAndRecordedListner = new("Benchmark.AllDataAndRecordedListner");
+        this.sourceWithOneProcessor = new("Benchmark.OneProcessor");
+        this.sourceWithTwoProcessors = new("Benchmark.TwoProcessors");
+        this.sourceWithThreeProcessors = new("Benchmark.ThreeProcessors");
+        this.sourceWithOneLegacyActivityOperationNameSubscription = new("Benchmark.OneInstrumentation");
+        this.sourceWithTwoLegacyActivityOperationNameSubscriptions = new("Benchmark.TwoInstrumentations");
+
+        this.activityListenerPropagationData = new ActivityListener
         {
             ActivityStarted = null,
             ActivityStopped = null,
             ShouldListenTo = (activitySource) => activitySource.Name == this.sourceWithPropagationDataListner.Name,
             Sample = (ref ActivityCreationOptions<ActivityContext> options) => ActivitySamplingResult.PropagationData,
-        });
+        };
 
-        ActivitySource.AddActivityListener(new ActivityListener
+        ActivitySource.AddActivityListener(this.activityListenerPropagationData);
+
+        this.activityListenerAllData = new ActivityListener
         {
             ActivityStarted = null,
             ActivityStopped = null,
             ShouldListenTo = (activitySource) => activitySource.Name == this.sourceWithAllDataListner.Name,
             Sample = (ref ActivityCreationOptions<ActivityContext> options) => ActivitySamplingResult.AllData,
-        });
+        };
 
-        ActivitySource.AddActivityListener(new ActivityListener
+        ActivitySource.AddActivityListener(this.activityListenerAllData);
+
+        this.activityListenerAllDataAndRecordedData = new ActivityListener
         {
             ActivityStarted = null,
             ActivityStopped = null,
             ShouldListenTo = (activitySource) => activitySource.Name == this.sourceWithAllDataAndRecordedListner.Name,
             Sample = (ref ActivityCreationOptions<ActivityContext> options) => ActivitySamplingResult.AllDataAndRecorded,
-        });
+        };
 
-        Sdk.CreateTracerProviderBuilder()
+        ActivitySource.AddActivityListener(this.activityListenerAllDataAndRecordedData);
+
+        this.tracerProvierWithOneProcessor = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddSource(this.sourceWithOneProcessor.Name)
+#pragma warning disable CA2000 // Dispose objects before losing scope
             .AddProcessor(new DummyActivityProcessor())
             .Build();
 
-        Sdk.CreateTracerProviderBuilder()
+        this.tracerProvierWithTwoProcessors = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddSource(this.sourceWithTwoProcessors.Name)
             .AddProcessor(new DummyActivityProcessor())
             .AddProcessor(new DummyActivityProcessor())
             .Build();
 
-        Sdk.CreateTracerProviderBuilder()
+        this.tracerProvierWithThreeProcessors = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddSource(this.sourceWithThreeProcessors.Name)
             .AddProcessor(new DummyActivityProcessor())
@@ -92,14 +124,14 @@ public class TraceBenchmarks
             .AddProcessor(new DummyActivityProcessor())
             .Build();
 
-        Sdk.CreateTracerProviderBuilder()
+        this.tracerProvierWithOneLegacyActivityOperationNameSubscription = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddSource(this.sourceWithOneLegacyActivityOperationNameSubscription.Name)
             .AddLegacySource("TestOperationName")
             .AddProcessor(new DummyActivityProcessor())
             .Build();
 
-        Sdk.CreateTracerProviderBuilder()
+        this.tracerProvierWithTwoLegacyActivityOperationNameSubscriptions = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddSource(this.sourceWithTwoLegacyActivityOperationNameSubscriptions.Name)
             .AddLegacySource("TestOperationName1")
@@ -107,90 +139,121 @@ public class TraceBenchmarks
             .AddProcessor(new DummyActivityProcessor())
             .Build();
 
-        Sdk.CreateTracerProviderBuilder()
+        this.tracerProvierWithExactMatchLegacyActivityListner = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddLegacySource("ExactMatch.OperationName1")
             .AddProcessor(new DummyActivityProcessor())
             .Build();
 
-        Sdk.CreateTracerProviderBuilder()
+        this.tracerProvierWithWildcardMatchLegacyActivityListner = Sdk.CreateTracerProviderBuilder()
             .SetSampler(new AlwaysOnSampler())
             .AddLegacySource("WildcardMatch.*")
             .AddProcessor(new DummyActivityProcessor())
+#pragma warning restore CA2000 // Dispose objects before losing scope
             .Build();
+    }
+
+    [GlobalCleanup]
+    public void Cleanup()
+    {
+        this.sourceWithNoListener?.Dispose();
+        this.sourceWithPropagationDataListner?.Dispose();
+        this.sourceWithAllDataListner?.Dispose();
+        this.sourceWithAllDataAndRecordedListner?.Dispose();
+        this.sourceWithOneProcessor?.Dispose();
+        this.sourceWithTwoProcessors?.Dispose();
+        this.sourceWithThreeProcessors?.Dispose();
+        this.sourceWithOneLegacyActivityOperationNameSubscription?.Dispose();
+        this.sourceWithTwoLegacyActivityOperationNameSubscriptions?.Dispose();
+
+        this.tracerProvierWithOneProcessor?.Dispose();
+        this.tracerProvierWithTwoProcessors?.Dispose();
+        this.tracerProvierWithThreeProcessors?.Dispose();
+        this.tracerProvierWithOneLegacyActivityOperationNameSubscription?.Dispose();
+        this.tracerProvierWithTwoLegacyActivityOperationNameSubscriptions?.Dispose();
+        this.tracerProvierWithExactMatchLegacyActivityListner?.Dispose();
+        this.tracerProvierWithWildcardMatchLegacyActivityListner?.Dispose();
+
+        this.activityListenerPropagationData?.Dispose();
+        this.activityListenerAllData?.Dispose();
+        this.activityListenerAllDataAndRecordedData?.Dispose();
     }
 
     [Benchmark]
     public void NoListener()
     {
         // this activity won't be created as there is no listener
-        using var activity = this.sourceWithNoListener.StartActivity("Benchmark");
+        using var activity = this.sourceWithNoListener!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void PropagationDataListner()
     {
         // this activity will be created and feed into an ActivityListener that simply drops everything on the floor
-        using var activity = this.sourceWithPropagationDataListner.StartActivity("Benchmark");
+        using var activity = this.sourceWithPropagationDataListner!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void AllDataListner()
     {
         // this activity will be created and feed into an ActivityListener that simply drops everything on the floor
-        using var activity = this.sourceWithAllDataListner.StartActivity("Benchmark");
+        using var activity = this.sourceWithAllDataListner!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void AllDataAndRecordedListner()
     {
         // this activity will be created and feed into an ActivityListener that simply drops everything on the floor
-        using var activity = this.sourceWithAllDataAndRecordedListner.StartActivity("Benchmark");
+        using var activity = this.sourceWithAllDataAndRecordedListner!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void OneProcessor()
     {
-        using var activity = this.sourceWithOneProcessor.StartActivity("Benchmark");
+        using var activity = this.sourceWithOneProcessor!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void TwoProcessors()
     {
-        using var activity = this.sourceWithTwoProcessors.StartActivity("Benchmark");
+        using var activity = this.sourceWithTwoProcessors!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void ThreeProcessors()
     {
-        using var activity = this.sourceWithThreeProcessors.StartActivity("Benchmark");
+        using var activity = this.sourceWithThreeProcessors!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void OneInstrumentation()
     {
-        using var activity = this.sourceWithOneLegacyActivityOperationNameSubscription.StartActivity("Benchmark");
+        using var activity = this.sourceWithOneLegacyActivityOperationNameSubscription!.StartActivity("Benchmark");
     }
 
     [Benchmark]
     public void TwoInstrumentations()
     {
-        using var activity = this.sourceWithTwoLegacyActivityOperationNameSubscriptions.StartActivity("Benchmark");
+        using var activity = this.sourceWithTwoLegacyActivityOperationNameSubscriptions!.StartActivity("Benchmark");
     }
 
     [Benchmark]
+#pragma warning disable CA1822 // Mark members as static
     public void LegacyActivity_ExactMatchMode()
+#pragma warning restore CA1822 // Mark members as static
     {
-        using var activity = new Activity("ExactMatch.OperationName1").Start();
+        using var activity = new Activity("ExactMatch.OperationName1");
+        activity.Start();
     }
 
     [Benchmark]
+#pragma warning disable CA1822 // Mark members as static
     public void LegacyActivity_WildcardMatchMode()
+#pragma warning restore CA1822 // Mark members as static
     {
-        using var activity = new Activity("WildcardMatch.OperationName1").Start();
+        using var activity = new Activity("WildcardMatch.OperationName1");
+        activity.Start();
     }
 
-    internal class DummyActivityProcessor : BaseProcessor<Activity>
-    {
-    }
+    internal sealed class DummyActivityProcessor : BaseProcessor<Activity>;
 }