+++ test/OpenTelemetry.Tests.Stress.Logs/Program.cs
@@ -5,16 +5,17 @@ using Microsoft.Extensions.Logging;
 
 namespace OpenTelemetry.Tests.Stress;
 
-public static class Program
+internal static class Program
 {
     public static int Main(string[] args)
     {
         return StressTestFactory.RunSynchronously<LogsStressTest>(args);
     }
 
+#pragma warning disable CA1812 // Avoid uninstantiated internal classes
     private sealed class LogsStressTest : StressTests<StressTestOptions>
+#pragma warning restore CA1812 // Avoid uninstantiated internal classes
     {
-        private static readonly Payload Payload = new();
         private readonly ILoggerFactory loggerFactory;
         private readonly ILogger logger;
 
@@ -32,10 +33,14 @@ public static class Program
             this.logger = this.loggerFactory.CreateLogger<LogsStressTest>();
         }
 
-        public override void Dispose()
+        protected override void Dispose(bool disposing)
         {
-            this.loggerFactory.Dispose();
-            base.Dispose();
+            if (disposing)
+            {
+                this.loggerFactory.Dispose();
+            }
+
+            base.Dispose(disposing);
         }
 
         protected override void RunWorkItemInParallel()